Move backends into own package
